Split script.js into a separate file for comments, fix style
