+++ portfolio/src/main/java/com/google/sps/data/Comment.java
@@ -21,12 +21,15 @@ public final class Comment {
   private final String name;
   private final String comment;
   private final long timestamp;
+  private final String email;
 
-  public Comment(long id, String name, String comment, long timestamp) {
+  public Comment(long id, String name, String comment, 
+    long timestamp, String email) {
 
     this.id = id;
     this.name = name;
     this.comment = comment;
     this.timestamp = timestamp;
+    this.email = email;
   }
 }
