+++ portfolio/src/main/webapp/forum.html
@@ -5,7 +5,7 @@
     <title>Forum</title>
     <link rel="stylesheet" href="style.css">
     <link rel="stylesheet" href="forum-style.css">
-    <script src="script.js"></script>
+    <script src="forum-script.js"></script>
   </head>
 
   <body onload="initComments()">
