+++ portfolio/src/main/webapp/index.html
@@ -7,7 +7,6 @@
     <script src="script.js"></script>
   </head>
 
-
   <body>
 
     <div class="nav-links">
